+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+- switched SNS to argus [#168](https://github.com/xmidt-org/tr1d1um/pull/168)
 
 ## [v0.5.1]
 ### Fixed